pyproject.toml
[project]
name = "sl-client-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "crewai>=1.7.2",
    "crewai-tools>=1.7.2",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "streamlit>=1.52.2",
    "openai>=1.0.0",
    "anthropic>=0.75.0",
    "fastapi>=0.128.0",
    "pydantic>=2.11.10",
]
